I have modified my menu screen to so that the  buttons now will deselect the endless mode option if they decide to choose the easy/normal/hard difficulty. I have decided to change this as this could possibly confuse the player what difficulty they are on. Furthermore it becomes then impossible to deselect the endless mode making players think they are playing on endless mode when they are not. I have also clarified in more details the amount of bullets you get. Which I have changed to meet the requirements of the client who said that more bullets should be given to the player otherwise, according to them, the game is too hard.
I have also modified the back menu to to say "Back To Main Menu" making it more clear to the user where the back button will take them. Although this is not a large change having small additions means that a wider range of people will be able to uderstand the game further and its built in mechanics.

I have also added my scores animation to be exponential (speeds up over time) to my end screen counter, instead of a linear progression. This makes the game feel more fluid as well as stops the game taking longer for the end screen to display each users final score. Making the game more fluid I feel will make a more enjoyable experience as they are not waiting a long time to have their score displayed.

When both the children and my friends tested the game they agreed the time it took to fire a bullet was too long. I have decreased this as you cannot shoot until 2 seconds has passed. Furthermore, I have changed the collisions to be more fair to the player, I was told by my friends it was too easy to get hit by the asteroids. I have now changed this to be more lenient to the player1 and player2. They will now only lose a life and collide with an asteroid if the asteroid is smaller than Y level 660.